requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.3
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-filter==24.1
django-oauth-toolkit==2.3.0
django-redis==5.4.0
django-rename-app==0.1.7
django-rest-framework==0.1.0
django-timezone-field==6.1.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-social-oauth2==2.1.3
drf-spectacular==0.27.1
ecdsa==0.18.0
frozenlist==1.4.1
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
kombu==5.3.5
multidict==6.0.5
oauthlib==3.2.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg==3.1.18
pyasn1==0.6.0
pycparser==2.22
pydot==2.0.0
PyJWT==2.8.0
pyparsing==3.1.2
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
python3-openid==3.2.0
PyYAML==6.0.1
redis==5.0.3
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
stripe==8.10.0
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
